chore(screenshot): update screenshot base class
